Initial Terraform type definition
This commit contains the first iteration of the
TerraformSpec and TerraformStatus types to be
used in the Terraform CRD.

This commit also imports the argo-cd package
and adds a patch to the go.mod to fix a bug
related to the k8s api not being imported
properly. There are comments in go.mod with more
information about this.

Signed-off-by: Joseph Sawaya <jsawaya@redhat.com>