Get terraform files from remote repo
This commit implements getting the terraform files
from the remote repository using the argocd-repo-server.

This commit implements an argocd plugin which is necessary
fetching the remote terraform files. This plugin calls a script
which runs in the path specified by the Terraform CR.

This commit creates a RepoClient to call the GenerateManifests
function in the repo server to make use of this plugin and get
the files. Finally it prints the contents of the manifests
returned by the the repo server.

Signed-off-by: Joseph Sawaya <jsawaya@redhat.com>