Get terraform files from remote repo (#16)
* Initial Terraform type definition

This commit contains the first iteration of the
TerraformSpec and TerraformStatus types to be
used in the Terraform CRD.

This commit also imports the argo-cd package
and adds a patch to the go.mod to fix a bug
related to the k8s api not being imported
properly. There are comments in go.mod with more
information about this.

Signed-off-by: Joseph Sawaya <jsawaya@redhat.com>

* upgrade argo pkg to v2 to fix build

Signed-off-by: Joseph Sawaya <jsawaya@redhat.com>

* update crd base

Signed-off-by: Joseph Sawaya <jsawaya@redhat.com>

* fix go.mod

Signed-off-by: Joseph <jsawaya@redhat.com>

* Get terraform files from remote repo

This commit implements getting the terraform files
from the remote repository using the argocd-repo-server.

This commit implements an argocd plugin which is necessary
fetching the remote terraform files. This plugin calls a script
which runs in the path specified by the Terraform CR.

This commit creates a RepoClient to call the GenerateManifests
function in the repo server to make use of this plugin and get
the files. Finally it prints the contents of the manifests
returned by the the repo server.

Signed-off-by: Joseph Sawaya <jsawaya@redhat.com>

* fix various config files

Signed-off-by: Joseph Sawaya <jsawaya@redhat.com>