Controller spawns worker pods (#17)
* Controller spawns worker pods

This commit updates the Terraform CR
definition so the spec only contain the
revision of the ArgoCD Application and
whether or not the workload has been completed
for this CR.

This commit also updates the terraform generate
script to generate a configmap alongside the
Terraform CR. The config map is created from
the directory the ArgoCD application is targetting.

This commit makes it so the controller now
spawns worker pods on reconciliation.

The worker pods run the argocd terraform
worker image which completes the terraform
workload. The config map generated by the
terraform-generate plugin is also mounted
to the worker pods.

This commit also mounts a secret as the env
in the worker pods to allow users to pass
their credentials through use of the secrets.

The worker pods' service account's roles are
limited to namespace only to allow for soft
multi-tenancy.

Signed-off-by: Joseph Sawaya <jsawaya@redhat.com>

* role file

Signed-off-by: Joseph Sawaya <jsawaya@redhat.com>

* config changes

Signed-off-by: Joseph Sawaya <jsawaya@redhat.com>

* Add worker code to this repository

This commit reorganizes the repository
and adds a Dockerfile to enable moving
the worker code here.

Signed-off-by: Joseph Sawaya <jsawaya@redhat.com>

* add kube namespace example

Signed-off-by: Joseph Sawaya <jsawaya@redhat.com>

* update kustomize for manager

Signed-off-by: Joseph Sawaya <jsawaya@redhat.com>

* Remove completed field from spec

This commit removes the completed field
from the spec since that is not a good
pattern for a kube operator. State was
being stored in the spec and that is not
idiomatic.

Signed-off-by: Joseph Sawaya <jsawaya@redhat.com>

* update readme

Signed-off-by: Joseph Sawaya <jsawaya@redhat.com>

* add default controller file

Signed-off-by: Joseph Sawaya <jsawaya@redhat.com>

* wip

Signed-off-by: Joseph Sawaya <jsawaya@redhat.com>